Amend erroneous check for Lua version ≥5.2.
`utils/load()` was checking whether it was running under Lua 5.2 or
later by checking whether the global variable `loadstring` was falsy.

I had evidently written this check under the impression that the
`loadstring` function was removed in Lua version 5.2, when it was in
actuality merely deprecated in that version.

This commit resolves the issue by checking for `setfenv` instead, that
function having indeed been removed in Lua version 5.2.

Incidentally, this was the first bug discovered by the new test suite.