Move `utils/dbgstr` to `butils.lua`.
Move `utils/dbgstr` to `butils.lua`, so that it can be called before the
`type` library is loaded, and alter its definition such that it can be
given a custom `typeof` function, which, if present, will be tried
before `type`.

In `utils.lua`, override `butils.lua`’s `dbgstr` with a wrapper function
that always passes `type/Type.of` as the custom `typeof` function.