Makefile: Allow for ignoring old server failures
In the `Makefile`, allow specifying a Boolean (`0` or `1`)
`ignore_failure_to_upload_to_old_addons_servers` parameter (default
`0`) as `1` to cause the script for the `release` target to ignore
failures to release to old add-ons servers (those that do not serve a
current BfW stable release or its corresponding development release,
if any).

At present, the old add-ons servers listed in the Makefile are those
for BfW versions 1.10.x and 1.11.x.