Change API macro namespace separator: ‘/’ → ‘:’
Change the character used as path component separator for namespaced
API macros from U+002F SOLIDUS (‘/’) to U+003A COLON (‘:’), because
vultraz thinks the solidi are confusing.

E.g., rename the `8680/lua-pack/path` macro to `8680:lua-pack:path`.

Include a compatibility shim for the only such macro that is
documented, `8680/lua-pack`, defining it now as equivalent to (and in
terms of) the new `8680:lua-pack`.