wml: Fix: Mark `get_subtags` for export
The `wml` module was supposed to export a `get_subtags`             
function, but did not, because the call of `lp8.export` that was  
intended to export `get_subtags` had been copied-and-pasted from  
the export statement for the `get_subtag` function, and not         
further modified (to refer to `get_subtags` rather than             
`get_subtag`).                                                      
                                                                    
This commit fixes this error by performing that modification.